Fix crash on Windows 7 systems
If the worker gets popped of the workers stack, Python does not
have any more references to the worker and the worker gets garbage
collected even if the thread has not finished. Assingning the worker
that has been poped of the stack to some dummy reference keeps the
worker in memory and prevents a native Qt crash where the worker gets
destroyed before finishing.